+++ src/components/molcules/Comment/index.stories.tsx
@@ -15,24 +15,24 @@ type Story = StoryObj<typeof meta>
 
 export const Primary: Story = {
   args: {
-    author: {
-      fullName: '사용자',
-      image: 'https://picsum.photos/200/200',
-    },
+    authorImage: 'https://picssum.phtos/200/200',
+    authorName: '사용자',
     comment: '그거 너무 싸게 산거임',
-    _id: '1',
     createAt: '2023-09-11',
   },
 }
 
 Primary.argTypes = {
-  _id: {
-    control: 'text',
-  },
   comment: {
     control: 'text',
   },
   createAt: {
+    control: 'text',
+  },
+  authorImage: {
+    control: 'text',
+  },
+  authorName: {
     control: 'text',
   },
 }
